furthers reorganization of model files, controllers, changes made to route and blade files
