+++ app/Models/Database/Permission_Types.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models\Database;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Permission_Types extends Model
+{
+    //
+}
