+++ app/Models/Database/Printers.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App;
+namespace App\Models\Database;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
 
-class Computers extends Model
+class Printers extends Model
 {
     //
 }
