+++ app/Models/Database/Subnets.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models\Database;
 
 use Illuminate\Database\Eloquent\Model;
