--- app/Sites.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Sites extends Model
-{
-    //
-}
-
